refactor(hydro_optimize,hydro_test): move optimizer tests into the optimizer crate
In preparation for splitting it out to a separate repo.
