+++ hydro_optimize/examples/benchmark_paxos.rs
