+++ hydro_optimize/examples/decouple_compute_pi.rs
