+++ hydro_optimize/examples/partition_simple_cluster.rs
