+++ hydro_optimize/examples/perf_compute_pi.rs
