+++ hydro_optimize/examples/perf_paxos.rs
